Fix IOS-190
Fix the missing @ symbol when there is no match.
Align the name and handler based on the center of the superview.
Fix the wrong radius corner for the profile picture.
Fix tableview insets when keyboard opens, now the scrollable area match what is seen on screen.
Add profile picture photo.